src/flask/requirements.txt
aniso8601==3.0.2
astroid==1.6.5
cassandra-driver==3.14.0
click==6.7
colorama==0.3.9
Flask==1.0.2
Flask-API==1.0
Flask-RESTful==0.3.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
pip==10.0.1
pylint==1.9.2
pytz==2018.5
scales==1.0.9
setuptools==39.0.1
six==1.11.0
Werkzeug==0.14.1
wheel==0.31.1
wrapt==1.10.11